+++ src/components/places-list-main/places-list-main.jsx
@@ -1,48 +1,34 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import PlaceCard from '../place-card/place-card.jsx';
+import {withActiveCardSwitcher} from '../../hocs/withActiveCardSwitcher/with-active-card-switcher.jsx';
 
-class PlacesListMain extends React.PureComponent {
-  constructor(props) {
-    super(props);
-    this.state = {
-      activeCardId: null
-    };
-    this.handleHover = this.handleHover.bind(this);
-  }
+const PlacesListMain = (props) => {
+  const {places, foundPlacesQnt, onPlaceCardClick, handleHover} = props;
 
-  handleHover(cardId) {
-    const value = (cardId === this.state.activeCardId) ? null : cardId;
-    this.setState({activeCardId: value});
-  }
-
-  render() {
-    const {places, foundPlacesQnt, onPlaceCardClick} = this.props;
-
-    return <section className="cities__places places">
-      <h2 className="visually-hidden">Places</h2>
-      <b className="places__found">{foundPlacesQnt} places to stay in Amsterdam</b>
-      <form className="places__sorting" action="#" method="get">
-        <span className="places__sorting-caption">Sort by</span>
-        <span className="places__sorting-type" tabIndex="0">
-          Popular
-          <svg className="places__sorting-arrow" width="7" height="4">
-            <use href="#icon-arrow-select"></use>
-          </svg>
-        </span>
-        <ul className="places__options places__options--custom places__options--opened">
-          <li className="places__option places__option--active" tabIndex="0">Popular</li>
-          <li className="places__option" tabIndex="0">Price: low to high</li>
-          <li className="places__option" tabIndex="0">Price: high to low</li>
-          <li className="places__option" tabIndex="0">Top rated first</li>
-        </ul>
-      </form>
-      <div className="cities__places-list places__list tabs__content">
-        {places.map((place, i) => <PlaceCard key={`key${i}`} cardId={`key${i}`} place={place} handleHover={this.handleHover} onPlaceCardClick={onPlaceCardClick} articleLocationClass={`cities`}/>)}
-      </div>
-    </section>;
-  }
-}
+  return <section className="cities__places places">
+    <h2 className="visually-hidden">Places</h2>
+    <b className="places__found">{foundPlacesQnt} places to stay in Amsterdam</b>
+    <form className="places__sorting" action="#" method="get">
+      <span className="places__sorting-caption">Sort by</span>
+      <span className="places__sorting-type" tabIndex="0">
+      Popular
+        <svg className="places__sorting-arrow" width="7" height="4">
+          <use href="#icon-arrow-select"></use>
+        </svg>
+      </span>
+      <ul className="places__options places__options--custom places__options--opened">
+        <li className="places__option places__option--active" tabIndex="0">Popular</li>
+        <li className="places__option" tabIndex="0">Price: low to high</li>
+        <li className="places__option" tabIndex="0">Price: high to low</li>
+        <li className="places__option" tabIndex="0">Top rated first</li>
+      </ul>
+    </form>
+    <div className="cities__places-list places__list tabs__content">
+      {places.map((place, i) => <PlaceCard key={`key${i}`} cardId={`key${i}`} place={place} handleHover={handleHover} onPlaceCardClick={onPlaceCardClick} articleLocationClass={`cities`}/>)}
+    </div>
+  </section>;
+};
 
 PlacesListMain.propTypes = {
   places: PropTypes.arrayOf(
@@ -60,7 +46,8 @@ PlacesListMain.propTypes = {
       }).isRequired
   ).isRequired,
   foundPlacesQnt: PropTypes.number.isRequired,
-  onPlaceCardClick: PropTypes.func.isRequired
+  onPlaceCardClick: PropTypes.func.isRequired,
+  handleHover: PropTypes.func.isRequired
 };
 
-export default PlacesListMain;
+export default withActiveCardSwitcher(PlacesListMain);
