+++ src/components/places-list-nearby/places-list-nearby.jsx
@@ -1,33 +1,18 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import PlaceCard from '../place-card/place-card.jsx';
+import {withActiveCardSwitcher} from '../../hocs/withActiveCardSwitcher/with-active-card-switcher.jsx';
 
-class PlacesListNearby extends React.PureComponent {
-  constructor(props) {
-    super(props);
-    this.state = {
-      activeCardId: null
-    };
-    this.handleHover = this.handleHover.bind(this);
-  }
+const PlacesListNearby = (props) => {
+  const {places, onPlaceCardClick, handleHover} = props;
 
-  handleHover(cardId) {
-    const value = (cardId === this.state.activeCardId) ? null : cardId;
-    this.setState({activeCardId: value});
-  }
-
-  render() {
-    const {places, onPlaceCardClick} = this.props;
-
-    return <section className="near-places places">
-      <h2 className="near-places__title">Other places in the neighbourhood</h2>
-      <div className="near-places__list places__list">
-        {places.map((place, i) => <PlaceCard key={`key${i}`} cardId={`key${i}`} place={place} handleHover={this.handleHover} onPlaceCardClick={onPlaceCardClick} articleLocationClass={`near-places`}/>)}
-      </div>
-    </section>;
-  }
-
-}
+  return <section className="near-places places">
+    <h2 className="near-places__title">Other places in the neighbourhood</h2>
+    <div className="near-places__list places__list">
+      {places.map((place, i) => <PlaceCard key={`key${i}`} cardId={`key${i}`} place={place} handleHover={handleHover} onPlaceCardClick={onPlaceCardClick} articleLocationClass={`near-places`}/>)}
+    </div>
+  </section>;
+};
 
 PlacesListNearby.propTypes = {
   places: PropTypes.arrayOf(
@@ -44,7 +29,8 @@ PlacesListNearby.propTypes = {
         }).isRequired
       }).isRequired
   ).isRequired,
-  onPlaceCardClick: PropTypes.func.isRequired
+  onPlaceCardClick: PropTypes.func.isRequired,
+  handleHover: PropTypes.func.isRequired
 };
 
-export default PlacesListNearby;
+export default withActiveCardSwitcher(PlacesListNearby);
